+++ common/src/main/java/top/vmctcn/vmtu/mod/modpack/info/ModpackInfoReader.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import top.vmctcn.vmtu.mod.ModContexts;
 import top.vmctcn.vmtu.mod.ModPlatform;
+import top.vmctcn.vmtu.mod.config.ModConfigHelper;
 
 import java.io.IOException;
 import java.io.Reader;
@@ -12,6 +13,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
+import java.util.Objects;
 
 public class ModpackInfoReader {
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
@@ -24,30 +26,33 @@ public class ModpackInfoReader {
             try (Reader reader = Files.newBufferedReader(modpackInfoPath, StandardCharsets.UTF_8)) {
                 modpackInfo = GSON.fromJson(reader, ModpackInfo.class);
                 if (modpackInfo == null) {
-                    ModContexts.LOGGER.warn("modpackinfo.json is empty or invalid, generating default file.");
-                    generateDefaultModpackInfo();
+                    if (ModConfigHelper.getConfig().misc.generateExampleModpackInfo == true) {
+                        ModContexts.LOGGER.warn("modpackinfo.json is empty or invalid, generating default file.");
+                        generateDefaultModpackInfo();
+                    } else {
+                        ModContexts.LOGGER.warn("modpackinfo.json is empty or invalid, skip it.");
+                    }
                 }
             } catch (Exception e) {
-                ModContexts.LOGGER.warn("Error reading modpackinfo.json, generating default file.", e);
-                generateDefaultModpackInfo();
+                if (ModConfigHelper.getConfig().misc.generateExampleModpackInfo == true) {
+                    ModContexts.LOGGER.warn("Error reading modpackinfo.json, generating default file.", e);
+                    generateDefaultModpackInfo();
+                } else {
+                    ModContexts.LOGGER.warn("Error reading modpackinfo.json, skip it.");
+                }
             }
         } else {
-            ModContexts.LOGGER.warn("modpackinfo.json does not exist, generating default file.");
-            generateDefaultModpackInfo();
+            if (ModConfigHelper.getConfig().misc.generateExampleModpackInfo == true) {
+                ModContexts.LOGGER.warn("modpackinfo.json does not exist, generating default file.");
+                generateDefaultModpackInfo();
+            } else {
+                ModContexts.LOGGER.warn("modpackinfo.json does not exist, skip it.");
+            }
         }
     }
 
     private static void generateDefaultModpackInfo() {
-        modpackInfo = new ModpackInfo();
-        modpackInfo.modpack = new ModpackInfo.Modpack();
-        modpackInfo.modpack.name = "ExampleModpack";
-        modpackInfo.modpack.version = "0.1.0";
-
-        modpackInfo.modpack.translation = new ModpackInfo.Translation();
-        modpackInfo.modpack.translation.id = "example";
-        modpackInfo.modpack.translation.url = "https://vmct-cn.top/modpacks/example/";
-        modpackInfo.modpack.translation.language = "zh_cn";
-        modpackInfo.modpack.translation.version = "1.0.0";
+        modpackInfo = new DefaultModpackInfo();
 
         try {
             try (Writer writer = Files.newBufferedWriter(modpackInfoPath, StandardCharsets.UTF_8, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING)) {
@@ -65,6 +70,10 @@ public class ModpackInfoReader {
     }
 
     public static ModpackInfo getModpackInfo() {
-        return modpackInfo;
+        return Objects.requireNonNullElseGet(modpackInfo, DefaultModpackInfo::new);
+    }
+
+    public static boolean isExampleModpackInfo() {
+        return getModpackInfo().getModpack().getTranslation().getId().equals("example");
     }
 }